requirements.txt
<<<<<<< HEAD
openai==1.64.0
python-dotenv==1.0.1
=======
crewai>=0.1.31
python-dotenv>=0.19.0 
>>>>>>> fe77e025f6fc03223534d5af71eb6244320f083b
